+++ src/errors.rs
@@ -114,6 +114,9 @@ pub enum IllFormedError {
     /// [specification]: https://www.w3.org/TR/xml11/#sec-comments
     /// [configuration]: crate::reader::Config::check_comments
     DoubleHyphenInComment,
+    /// The parser started to parse entity or character reference (`&...;`) in text,
+    /// but the input ended before the closing `;` character was found.
+    UnclosedReference,
 }
 
 impl fmt::Display for IllFormedError {
@@ -144,6 +147,9 @@ impl fmt::Display for IllFormedError {
             Self::DoubleHyphenInComment => {
                 f.write_str("forbidden string `--` was found in a comment")
             }
+            Self::UnclosedReference => f.write_str(
+                "entity or character reference not closed: `;` not found before end of input",
+            ),
         }
     }
 }