+++ Changelog.md
@@ -13,11 +13,20 @@
 
 ## Unreleased
 
+### Significant changes
+
+Now references to entities (as predefined, such as `&lt;`, as user-defined) reported as a new
+`Event::GeneralRef`.
+Caller can parse the content of the entity and stream events from it as it is required by the
+XML specification.
+
 ### New Features
 
 - [#863]: Add `Attributes::into_map_access(&str)` and `Attributes::into_deserializer()` when `serialize`
   feature is enabled. This will allow do deserialize serde types right from attributes. Both methods
   returns the same type which implements serde's `Deserializer` and `MapAccess` traits.
+- [#766]: Allow to parse resolved entities as XML fragments and stream events from them.
+- [#766]: Added new event `Event::GeneralRef` with content of [general entity].
 
 ### Bug Fixes
 
@@ -26,7 +35,9 @@
 - [#863]: Remove `From<QName<'a>> for BytesStart<'a>` because now `BytesStart` stores the
   encoding in which its data is encoded, but `QName` is a simple wrapper around byte slice.
 
+[#766]: https://github.com/tafia/quick-xml/pull/766
 [#863]: https://github.com/tafia/quick-xml/pull/863
+[general entity]: https://www.w3.org/TR/xml11/#gen-entity
 
 
 ## 0.37.5 -- 2025-04-27